+++ mono_core/lib/src/config/config_interfaces.dart
@@ -4,15 +4,3 @@ abstract class ConfigLoader {
   const ConfigLoader();
   MonoConfig load(String text);
 }
-
-@immutable
-abstract class ConfigValidator {
-  const ConfigValidator();
-  List<ConfigIssue> validate(MonoConfig config);
-}
-
-@immutable
-abstract class SchemaProvider {
-  const SchemaProvider();
-  Map<String, Object?> jsonSchema();
-}